requirements.txt
attrs==21.2.0
base58==2.1.0
# Editable Git install with no remote (indyctl==0.1.0)
-e /home/mrk/devel/indy/indyctl
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python3-indy==1.16.0
PyYAML==5.4.1
toml==0.10.2
